// Stage Drive/Dashboard/justbitteam-stagedriving_dash-4cdbe67ac041/src/app/api/model/accountDTO.ts
/**
 * Stage Driving API
 * API specifications
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AccountDeviceDTO } from './accountDeviceDTO';
import { AccountGroupDTO } from './accountGroupDTO';
import { AccountImageDTO } from './accountImageDTO';
import { AccountMetaDTO } from './accountMetaDTO';
import { ConnectionDTO } from './connectionDTO';
import { EventDTO } from './eventDTO';
import { VehicleDTO } from './vehicleDTO';


/**
 * Account model
 */
export interface AccountDTO {
    id?: string;
    created?: string;
    modified?: string;
    firstname?: string;
    middlename?: string;
    lastname?: string;
    token?: string;
    email: string;
    address?: string;
    country?: string;
    town?: string;
    city?: string;
    zipcode?: string;
    status?: string;
    gender?: string;
    birthday?: Date;
    telephone?: string;
    mobile?: string;
    pec?: string;
    note?: string;
    rating?: number;
    favStyle?: string;
    favCategory?: string;
    password?: string;
    expires?: Date;
    role?: string;
    username?: string;
    visible?: boolean;
    active?: boolean;
    isFriend?: boolean;
    companyVatId?: string;
    companyName?: string;
    companyAddress?: string;
    companyCountry?: string;
    companyCity?: string;
    companyZipcode?: string;
    companyRef?: string;
    vehicles?: Array<VehicleDTO>;
    devices?: Array<AccountDeviceDTO>;
    metas?: Array<AccountMetaDTO>;
    images?: Array<AccountImageDTO>;
    connections?: Array<ConnectionDTO>;
    groups?: Array<AccountGroupDTO>;
    event?: EventDTO;
}